Avoid BigDecimal completely

+++ src/main/java/it/polimi/diceH2020/SPACE4CloudWS/solvers/Solver.java
@@ -19,13 +19,9 @@ package it.polimi.diceH2020.SPACE4CloudWS.solvers;
 import it.polimi.diceH2020.SPACE4Cloud.shared.solution.SolutionPerJob;
 import it.polimi.diceH2020.SPACE4CloudWS.services.SshConnectorProxy;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
-/**
- * Created by ciavotta on 11/02/16.
- */
 public interface Solver {
 
     void setAccuracy(double accuracy);
@@ -38,7 +34,7 @@ public interface Solver {
 
     SshConnectorProxy getConnector();
 
-    Optional<BigDecimal> evaluate(SolutionPerJob solPerJob);
+    Optional<Double> evaluate(SolutionPerJob solPerJob);
 
     void restoreDefaults();
